TST: reorganize IO tests
